requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.4.1
billiard==3.6.3.0
bugsnag==4.0.1
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.6
django-cachalot==2.2.0
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-debug-toolbar-request-history==0.1.3
django-filter==2.2.0
django-redis==4.11.0
djangorestframework==3.11.0
drf-yasg==1.20.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jsonpickle==1.4.1
kiwisolver==1.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
packaging==20.4
pandas==1.0.3
plotly==4.7.1
psycopg2==2.8.5
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.2
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
WebOb==1.8.6
wrapt==1.12.1
zipp==3.1.0
